Added DBConnection.php a simple page to hold the PDO object.
Added Database_error.php a simple error page if the DB connection does not work.

Added a Navbar that show links on the left side of the page to all of the pages in the chart.

Added a Patient.php class that allows us to pass a Patient object around. need way more work.!

Added a patientIDBar that is at the top of each page and show the users info.

Added a simple index test page to test the layout of the nav bar and the patient IDbar